Allow broadcast requests to the servers
It allows the client to use requests of the form:
"coap://255.255.255.255/.well-known/core" to search for primitive
servers on the local area network (LAN) that do not support multicasts.

It also enables the client to reuse sockets, since the packets have
a built-in mechanism for determining if the response matches the request.